+++ HearthWatcher/Providers/IDiscoverStateProvider.cs
@@ -1,0 +1,8 @@
+﻿using HearthMirror.Objects;
+
+namespace HearthWatcher.Providers;
+
+public interface IDiscoverStateProvider
+{
+	DiscoverState? State { get; }
+}